remove private module
