--- src/tahrir_api/_compat.py
@@ -1,19 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-.. $Id$
-"""
-
-from __future__ import division
-from __future__ import print_function
-from __future__ import absolute_import
-
-
-def bytes_(s, encoding='utf-8', errors='strict'):
-    """
-    If ``s`` is an instance of ``text_type``, return
-    ``s.encode(encoding, errors)``, otherwise return ``s``
-    """
-    if not isinstance(s, bytes) and s is not None:
-        s = s.encode(encoding, errors)
-    return s
