+++ src/tahrir_api/utils.py
@@ -41,3 +41,13 @@ def convert_name_to_id(name):
     for a, b in replacements.items():
         badge_id = badge_id.replace(a, b)
     return badge_id
+
+
+def bytes_(s, encoding='utf-8', errors='strict'):
+    """
+    If ``s`` is an instance of ``text_type``, return
+    ``s.encode(encoding, errors)``, otherwise return ``s``
+    """
+    if not isinstance(s, bytes) and s is not None:
+        s = s.encode(encoding, errors)
+    return s
